pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pheasant"
version = "3.0.0"
description = "Documentation tool for Markdown conversion by Jupyter client"
readme = "README.md"
license = "MIT"
authors = [{ name = "daizutabi", email = "daizutabi@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = ["mkdocs>=1.6", "ipykernel>=6"]

[project.urls]
Documentation = "https://github.com/daizutabi/pheasant"
Source = "https://github.com/daizutabi/pheasant"
Issues = "https://github.com/daizutabi/pheasant/issues"

[dependency-groups]
dev = [
  "mkdocs-material",
  "pytest-clarity>=1",
  "pytest-cov>=6",
  "pytest-randomly>=3.16",
  "pytest-xdist>=3.6",
]

[tool.pytest.ini_options]
addopts = ["--doctest-modules", "--cov=pheasant", "--cov-report=lcov:lcov.info"]

[tool.coverage.report]
exclude_lines = ["no cov", "raise NotImplementedError", "if TYPE_CHECKING:"]
skip_covered = true

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
unfixable = ["F401"]
ignore = [
  "ANN003",
  "ANN401",
  "ARG002",
  "B904",
  "D105",
  "D107",
  "D203",
  "D213",
  "EM101",
  "PGH003",
  "TRY003",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "A001",
  "ANN",
  "ARG",
  "D",
  "FBT",
  "PD",
  "PLR",
  "PT",
  "S",
  "SIM117",
  "SLF",
]

[tool.pyright]
include = ["src", "tests"]
strictDictionaryInference = true
strictListInference = true
strictSetInference = true
